Do a proper snapshot instead of dumb copy of workspace
